# Improve hash_si performance for serializing strings, fix php 5 test failure.