serialize_object_array sped up from 2.27 to 1.85 seconds.
Minor improvements of serialize_string_array (0.20s to 0.18s)

Fix typo in earlier test.